WikiDom: added `TypeDict` data structure.
The values stored in this dictionary are indexed by their //class//,
which means that for a given key, the value stored for the nearest
corresponding ancestor will be returned.

In other words, this provides the same behavior as one would get
by adding class attributes to a class hierarchy, but without
actually modifying those classes.