WikiDom: slight detour by the formatting side of things.
To be sure we're on the right track when building the parser, we'll
try to jointly check the we can effectively render the parse tree
we're building.

We're therefore adding the new page, inline and outline formatters,
even if they don't do much for now, as the parse tree is embryonic at
this stage.